Adds a notification model and tests
Adds a model so that users can express a preference wrt receiving issues notifications.
The assumed logic is that when checking, if no record is found then whether they receive
the notification or not is determined by configuration defaults (upcoming commit).

1. If a user chooses 'never', they will never receive an issue notification

2. If the user chooses 'all_publishers' then they will receive a notification
unless that publisher is specified in 'exclude_publishers'

3. If user explicitly includes/excludes a publisher then that value will be
used.

4. Else the all_where_editor_admin property is used to determine the
answer.

Includes NotificationTokens for quick 'login' to subscription prefs.

Contains a table (and tests) that records a code against a user id
allowing them to change their preferences (but nothing else). This does
not log them in, it just allows them to change the settings for emails.

Adds outline of the controller action

Adds a controller action and route for managing subscriptions, but is incomplete.

Logic should be moved to logic_layer once it is working.